fix(): Handle NSC pods with no parent containers
Some container runtimes might not have a parent container
in kubernetes pods like the pause or the podman container.
In such environments, the mapping between the network namespace inode
to the pid of the parent container cannot be maintained. Instead, we
end up using the pid of the nsc-init container in the mapping, but the
map entry becomes invalid once the init container exits and the rest
of the containers in the pod start up.
To fix this problem, we will rewrite the mapping with the pid of a
regular container in the pod.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>